Drop PHP 7.2 support
