--- src/test/resources/files/Source/Language/PHP/PHPParserVersion72/testTrailingCommasInUnsetCall.php
@@ -1,5 +1,0 @@
-<?php
-
-function remove(&$i) {
-    unset($i,);
-}
